refactor: PR #344 Round 4 review 코멘트 반영 (스킬 통계 이동, 중복 제거, 마법 상수 제거, useCallback 추가)
